+++ zarrs_codec/src/codec_partial_default.rs
@@ -162,7 +162,8 @@ where
             Ok(match self.decoded_representation.data_type().size() {
                 DataTypeSize::Fixed(_) => ArrayBytes::new_flen(vec![]),
                 DataTypeSize::Variable => {
-                    ArrayBytes::new_vlen(vec![], ArrayBytesOffsets::new(vec![0]).unwrap()).unwrap()
+                    ArrayBytes::new_vlen(vec![], ArrayBytesOffsets::new(vec![0u32]).unwrap())
+                        .unwrap()
                 }
             })
         }
@@ -516,7 +517,8 @@ where
             Ok(match self.decoded_representation.data_type().size() {
                 DataTypeSize::Fixed(_) => ArrayBytes::new_flen(vec![]),
                 DataTypeSize::Variable => {
-                    ArrayBytes::new_vlen(vec![], ArrayBytesOffsets::new(vec![0]).unwrap()).unwrap()
+                    ArrayBytes::new_vlen(vec![], ArrayBytesOffsets::new(vec![0u32]).unwrap())
+                        .unwrap()
                 }
             })
         }